+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/AppNamespaceService.java
@@ -27,14 +27,13 @@ import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
-
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
 
 @Service
 public class AppNamespaceService {
@@ -254,4 +253,5 @@ public class AppNamespaceService {
   public void batchDeleteByAppId(String appId, String operator) {
     appNamespaceRepository.batchDeleteByAppId(appId, operator);
   }
+
 }