+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/NamespaceService.java
@@ -17,6 +17,7 @@
 package com.ctrip.framework.apollo.portal.service;
 
 import com.ctrip.framework.apollo.common.constants.GsonType;
+import com.ctrip.framework.apollo.common.dto.ClusterDTO;
 import com.ctrip.framework.apollo.common.dto.ItemDTO;
 import com.ctrip.framework.apollo.common.dto.NamespaceDTO;
 import com.ctrip.framework.apollo.common.dto.PageDTO;
@@ -28,6 +29,7 @@ import com.ctrip.framework.apollo.core.enums.ConfigFileFormat;
 import com.ctrip.framework.apollo.core.utils.ApolloThreadFactory;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
 import com.ctrip.framework.apollo.portal.api.AdminServiceAPI;
+import com.ctrip.framework.apollo.portal.api.AdminServiceAPI.NamespaceAPI;
 import com.ctrip.framework.apollo.portal.component.PortalSettings;
 import com.ctrip.framework.apollo.portal.component.config.PortalConfig;
 import com.ctrip.framework.apollo.portal.constant.RoleType;
@@ -35,6 +37,7 @@ import com.ctrip.framework.apollo.portal.constant.TracerEventType;
 import com.ctrip.framework.apollo.portal.enricher.adapter.BaseDtoUserInfoEnrichedAdapter;
 import com.ctrip.framework.apollo.portal.entity.bo.ItemBO;
 import com.ctrip.framework.apollo.portal.entity.bo.NamespaceBO;
+import com.ctrip.framework.apollo.portal.entity.vo.NamespaceUsage;
 import com.ctrip.framework.apollo.portal.environment.Env;
 import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
 import com.ctrip.framework.apollo.portal.util.RoleUtils;
@@ -42,6 +45,7 @@ import com.ctrip.framework.apollo.tracer.Tracer;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.gson.Gson;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -80,19 +84,21 @@ public class NamespaceService {
   private final NamespaceBranchService branchService;
   private final RolePermissionService rolePermissionService;
   private final AdditionalUserInfoEnrichService additionalUserInfoEnrichService;
+  private final ClusterService clusterService;
 
   public NamespaceService(
       final PortalConfig portalConfig,
       final PortalSettings portalSettings,
       final UserInfoHolder userInfoHolder,
-      final AdminServiceAPI.NamespaceAPI namespaceAPI,
+      final NamespaceAPI namespaceAPI,
       final ItemService itemService,
       final ReleaseService releaseService,
       final AppNamespaceService appNamespaceService,
       final InstanceService instanceService,
       final @Lazy NamespaceBranchService branchService,
       final RolePermissionService rolePermissionService,
-      final AdditionalUserInfoEnrichService additionalUserInfoEnrichService) {
+      final AdditionalUserInfoEnrichService additionalUserInfoEnrichService,
+      ClusterService clusterService) {
     this.portalConfig = portalConfig;
     this.portalSettings = portalSettings;
     this.userInfoHolder = userInfoHolder;
@@ -104,6 +110,7 @@ public class NamespaceService {
     this.branchService = branchService;
     this.rolePermissionService = rolePermissionService;
     this.additionalUserInfoEnrichService = additionalUserInfoEnrichService;
+    this.clusterService = clusterService;
   }
 
 
@@ -124,35 +131,46 @@ public class NamespaceService {
   }
 
 
+  public List<NamespaceUsage> getNamespaceUsageByAppId(String appId, String namespaceName) {
+    List<Env> envs = portalSettings.getActiveEnvs();
+    AppNamespace appNamespace = appNamespaceService.findByAppIdAndName(appId, namespaceName);
+    List<NamespaceUsage> usages = new ArrayList<>();
+    for (Env env : envs) {
+      List<ClusterDTO> clusters = clusterService.findClusters(env, appId);
+      for (ClusterDTO cluster : clusters) {
+        String clusterName = cluster.getName();
+        NamespaceUsage usage = this.getNamespaceUsageByEnv(appId, namespaceName, env, clusterName);
+        if (appNamespace != null && appNamespace.isPublic()) {
+          int associatedNamespace = this.getPublicAppNamespaceHasAssociatedNamespace(namespaceName, env);
+          usage.setLinkedNamespaceCount(associatedNamespace);
+        }
+
+        if(usage.getLinkedNamespaceCount() > 0 || usage.getBranchInstanceCount() > 0 || usage.getInstanceCount() > 0) {
+          usages.add(usage);
+        }
+      }
+    }
+    return usages;
+  }
+
+  public NamespaceUsage getNamespaceUsageByEnv(String appId, String namespaceName, Env env, String clusterName) {
+    NamespaceUsage namespaceUsage = new NamespaceUsage(namespaceName, appId, clusterName, env.getName());
+    int instanceCount = instanceService.getInstanceCountByNamespace(appId, env, clusterName, namespaceName);
+    namespaceUsage.setInstanceCount(instanceCount);
+
+    NamespaceDTO branchNamespace = branchService.findBranchBaseInfo(appId, env, clusterName, namespaceName);
+    if(branchNamespace != null){
+      String branchClusterName = branchNamespace.getClusterName();
+      int branchInstanceCount = instanceService.getInstanceCountByNamespace(appId, env, branchClusterName, namespaceName);
+      namespaceUsage.setBranchInstanceCount(branchInstanceCount);
+    }
+    return namespaceUsage;
+  }
+
   @Transactional
   public void deleteNamespace(String appId, Env env, String clusterName, String namespaceName) {
 
-    AppNamespace appNamespace = appNamespaceService.findByAppIdAndName(appId, namespaceName);
-
-    //1. check parent namespace has not instances
-    if (namespaceHasInstances(appId, env, clusterName, namespaceName)) {
-      throw new BadRequestException(
-          "Can not delete namespace because namespace has active instances");
-    }
-
-    //2. check child namespace has not instances
-    NamespaceDTO childNamespace = branchService
-        .findBranchBaseInfo(appId, env, clusterName, namespaceName);
-    if (childNamespace != null &&
-        namespaceHasInstances(appId, env, childNamespace.getClusterName(), namespaceName)) {
-      throw new BadRequestException(
-          "Can not delete namespace because namespace's branch has active instances");
-    }
-
-    //3. check public namespace has not associated namespace
-    if (appNamespace != null && appNamespace.isPublic() && publicAppNamespaceHasAssociatedNamespace(
-        namespaceName, env)) {
-      throw new BadRequestException(
-          "Can not delete public namespace which has associated namespaces");
-    }
-
     String operator = userInfoHolder.getUser().getUserId();
-
     namespaceAPI.deleteNamespace(env, appId, clusterName, namespaceName, operator);
   }
 
@@ -236,13 +254,12 @@ public class NamespaceService {
     return transformNamespace2BO(env, namespace);
   }
 
-  public boolean namespaceHasInstances(String appId, Env env, String clusterName,
-      String namespaceName) {
-    return instanceService.getInstanceCountByNamespace(appId, env, clusterName, namespaceName) > 0;
+  public boolean publicAppNamespaceHasAssociatedNamespace(String publicNamespaceName, Env env) {
+    return getPublicAppNamespaceHasAssociatedNamespace(publicNamespaceName, env) > 0;
   }
 
-  public boolean publicAppNamespaceHasAssociatedNamespace(String publicNamespaceName, Env env) {
-    return namespaceAPI.countPublicAppNamespaceAssociatedNamespaces(env, publicNamespaceName) > 0;
+  public int getPublicAppNamespaceHasAssociatedNamespace(String publicNamespaceName, Env env) {
+    return namespaceAPI.countPublicAppNamespaceAssociatedNamespaces(env, publicNamespaceName);
   }
 
   public NamespaceBO findPublicNamespaceForAssociatedNamespace(Env env, String appId,