+++ apollo-portal/src/main/resources/static/scripts/controller/config/ConfigNamespaceController.js
@@ -404,33 +404,6 @@ function controller($rootScope, $scope, $translate, toastr, AppUtil, EventManage
 
     }
 
-    EventManager.subscribe(EventManager.EventType.DELETE_NAMESPACE_FAILED, function (context) {
-        $scope.deleteNamespaceContext = context;
-
-        if (context.reason == 'master_instance') {
-            AppUtil.showModal('#deleteNamespaceDenyForMasterInstanceDialog');
-        } else if (context.reason == 'branch_instance') {
-            AppUtil.showModal('#deleteNamespaceDenyForBranchInstanceDialog');
-        } else if (context.reason == 'public_namespace') {
-            var otherAppAssociatedNamespaces = context.otherAppAssociatedNamespaces;
-            var namespaceTips = [];
-            otherAppAssociatedNamespaces.forEach(function (namespace) {
-                var appId = namespace.appId;
-                var clusterName = namespace.clusterName;
-                var url = AppUtil.prefixPath() + '/config.html?#/appid=' + appId + '&env=' + $scope.pageContext.env + '&cluster='
-                    + clusterName;
-
-                namespaceTips.push("<a target='_blank' href=\'" + url + "\'>AppId = " + appId + ", Cluster = " + clusterName
-                    + ", Namespace = " + namespace.namespaceName + "</a>");
-            });
-
-            $scope.deleteNamespaceContext.detailReason = $translate.instant('Config.DeleteNamespaceFailedTips') + "<br>" + namespaceTips.join("<br>");
-
-            AppUtil.showModal('#deleteNamespaceDenyForPublicNamespaceDialog');
-        }
-
-    });
-
     EventManager.subscribe(EventManager.EventType.SYNTAX_CHECK_TEXT_FAILED, function (context) {
         $scope.syntaxCheckContext = context;