+++ apollo-portal/src/main/resources/static/scripts/directive/delete-namespace-modal-directive.js
@@ -28,8 +28,6 @@ function deleteNamespaceModalDirective($window, $q, $translate, toastr, AppUtil,
         },
         link: function (scope) {
 
-
-
             scope.doDeleteNamespace = doDeleteNamespace;
 
             EventManager.subscribe(EventManager.EventType.PRE_DELETE_NAMESPACE, function (context) {
@@ -39,21 +37,25 @@ function deleteNamespaceModalDirective($window, $q, $translate, toastr, AppUtil,
                 //1. check operator has master permission
                 checkPermission(toDeleteNamespace).then(function () {
 
-                    //2. check namespace's master branch has not instances
-                    if (!checkMasterInstance(toDeleteNamespace)) {
-                        return;
-                    }
-
-                    //3. check namespace's gray branch has not instances
-                    if (!checkBranchInstance(toDeleteNamespace)) {
-                        return;
-                    }
-
-                    if (!toDeleteNamespace.isPublic || toDeleteNamespace.isLinkedNamespace) {
-                        showDeleteNamespaceConfirmDialog();
+                    if (toDeleteNamespace.isLinkedNamespace) {
+                        NamespaceService.getLinkedNamespaceUsage(toDeleteNamespace.baseInfo.appId, scope.env,
+                            toDeleteNamespace.baseInfo.clusterName,
+                            toDeleteNamespace.baseInfo.namespaceName
+                        ).then(function (usage) {
+                            scope.toDeleteNamespace.namespaceUsage = usage;
+                            if (usage[0].instanceCount > 0 || usage[0].branchInstanceCount > 0) {
+                                scope.toDeleteNamespace.forceDeleteButton = true;
+                            }
+                            showDeleteNamespaceConfirmDialog();
+                        });
                     } else {
-                        //5. check public namespace has not associated namespace
-                        checkPublicNamespace(toDeleteNamespace).then(function () {
+                        NamespaceService.getNamespaceUsage(toDeleteNamespace.baseInfo.appId,
+                            toDeleteNamespace.baseInfo.namespaceName
+                        ).then(function (usage) {
+                            scope.toDeleteNamespace.namespaceUsage = usage;
+                            if (usage.length > 0) {
+                                scope.toDeleteNamespace.forceDeleteButton = true;
+                            }
                             showDeleteNamespaceConfirmDialog();
                         });
                     }
@@ -98,73 +100,16 @@ function deleteNamespaceModalDirective($window, $q, $translate, toastr, AppUtil,
                 return d.promise;
             }
 
-            function checkMasterInstance(namespace) {
-                if (namespace.instancesCount > 0) {
-                    EventManager.emit(EventManager.EventType.DELETE_NAMESPACE_FAILED, {
-                        namespace: namespace,
-                        reason: 'master_instance'
-                    });
-
-                    return false;
-                }
-
-                return true;
-            }
-
-            function checkBranchInstance(namespace) {
-                if (namespace.hasBranch && namespace.branch.latestReleaseInstances.total > 0) {
-                    EventManager.emit(EventManager.EventType.DELETE_NAMESPACE_FAILED, {
-                        namespace: namespace,
-                        reason: 'branch_instance'
-                    });
-
-                    return false;
-                }
-
-                return true;
-            }
-
-            function checkPublicNamespace(namespace) {
-                var d = $q.defer();
-
-                var publicAppId = namespace.baseInfo.appId;
-                NamespaceService.getPublicAppNamespaceAllNamespaces(scope.env,
-                    namespace.baseInfo.namespaceName,
-                    0, 20)
-                    .then(function (associatedNamespaces) {
-                        var otherAppAssociatedNamespaces = [];
-                        associatedNamespaces.forEach(function (associatedNamespace) {
-                            if (associatedNamespace.appId != publicAppId) {
-                                otherAppAssociatedNamespaces.push(associatedNamespace);
-                            }
-                        });
-
-                        if (otherAppAssociatedNamespaces.length) {
-                            EventManager.emit(EventManager.EventType.DELETE_NAMESPACE_FAILED, {
-                                namespace: namespace,
-                                reason: 'public_namespace',
-                                otherAppAssociatedNamespaces: otherAppAssociatedNamespaces
-                            });
-                            d.reject();
-                        } else {
-                            d.resolve();
-                        }
-
-                    });
-
-                return d.promise;
-
-            }
-
             function showDeleteNamespaceConfirmDialog() {
                 AppUtil.showModal('#deleteNamespaceModal');
             }
 
             function doDeleteNamespace() {
                 var toDeleteNamespace = scope.toDeleteNamespace;
-                NamespaceService.deleteNamespace(toDeleteNamespace.baseInfo.appId, scope.env,
-                    toDeleteNamespace.baseInfo.clusterName,
-                    toDeleteNamespace.baseInfo.namespaceName)
+                if(toDeleteNamespace.isLinkedNamespace){
+                    NamespaceService.deleteLinkedNamespace(toDeleteNamespace.baseInfo.appId, scope.env,
+                        toDeleteNamespace.baseInfo.clusterName,
+                        toDeleteNamespace.baseInfo.namespaceName)
                     .then(function () {
                         toastr.success($translate.instant('Common.Deleted'));
 
@@ -175,7 +120,20 @@ function deleteNamespaceModalDirective($window, $q, $translate, toastr, AppUtil,
                     }, function (result) {
                         AppUtil.showErrorMsg(result, $translate.instant('Common.DeleteFailed'));
                     })
+                } else {
+                    NamespaceService.deleteAppNamespace(toDeleteNamespace.baseInfo.appId,
+                        toDeleteNamespace.baseInfo.namespaceName)
+                    .then(function () {
+                        toastr.success($translate.instant('Common.Deleted'));
 
+                        setTimeout(function () {
+                            $window.location.reload();
+                        }, 1000);
+
+                    }, function (result) {
+                        AppUtil.showErrorMsg(result, $translate.instant('Common.DeleteFailed'));
+                    })
+                }
             }
 
         }