+++ apollo-portal/src/main/resources/static/scripts/services/EventManager.js
@@ -127,7 +127,6 @@ appService.service('EventManager', [function () {
             PRE_DELETE_NAMESPACE: 'pre_delete_namespace',
             PRE_IMPORT_NAMESPACE: 'pre_import_namespace',
             DELETE_NAMESPACE: 'delete_namespace',
-            DELETE_NAMESPACE_FAILED: 'delete_namespace_failed',
             CHANGE_ENV_CLUSTER: "change_env_cluster",
             SYNTAX_CHECK_TEXT_FAILED: "syntax_check_text_failed"
         }