+++ CHANGES.md
@@ -23,6 +23,7 @@ Apollo 2.1.0
 * [fix(#4474):'openjdk:8-jre-alpine' potentially causing wrong number of cpu cores](https://github.com/apolloconfig/apollo/pull/4475)
 * [Switching spring-session serialization mode to json for compatibility with spring-security version updates]()
 * [fix(#4483):Fixed overwrite JSON type configuration being empty](https://github.com/apolloconfig/apollo/pull/4486)
+* [Allow users to delete AppNamespace](https://github.com/apolloconfig/apollo/pull/4499)
 * [fix the deleted at timestamp issue](https://github.com/apolloconfig/apollo/pull/4493)
 
 ------------------